correcting issues with "'"

+++ piper-tts.py
@@ -17,19 +17,22 @@ def notify(text):
 
 def read_aloud_stream(text):
     notify('Reading aloud')
-    text = text.replace("'", "`")
-    command = f"echo '{text}' |  {piper_dir} --model {directory}/{voice_model} -s 7 --output-raw | aplay -r 22050 -f S16_LE -t raw -"
-    return subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+    echo_process = subprocess.Popen(["echo", text], stdout=subprocess.PIPE)
+    piper_process = subprocess.Popen([piper_dir, "--model", directory + "/" + voice_model, "-s", "7", "--output-raw"], stdin=echo_process.stdout, stdout=subprocess.PIPE)
+    aplay_process = subprocess.Popen(["aplay", "-r", "22050", "-f", "S16_LE", "-t", "raw", "-"], stdin=piper_process.stdout, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+    echo_process.stdout.close()
+    piper_process.stdout.close()
+    return aplay_process
 
 def read_aloud_file(text):
     notify('Reading aloud, press CTRL+ALT to stop')
-    text = text.replace("'", "`")
-    command = f"echo '{text}' | {piper_dir} --model {directory}/{voice_model} -s 7 --output_file {directory}/output.wav; open {directory}/output.wav"
-    print(command)
-    subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    print('done')
-    time.sleep(0.01)
-
+    echo_process = subprocess.Popen(["echo", text], stdout=subprocess.PIPE)
+    piper_process = subprocess.Popen([piper_dir, "--model", directory + "/" + voice_model, "-s", "7", "--output_file", directory+'/output.wav'], stdin=echo_process.stdout, stdout=subprocess.PIPE)
+    time.sleep(0.1)
+    aplay_process = subprocess.Popen(["open",directory+"/output.wav"], stdin=piper_process.stdout, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+    echo_process.stdout.close()
+    piper_process.stdout.close()
+    aplay_process.stdout.close()
 
 def on_press(key):
     pressed_keys.add(key)
@@ -46,14 +49,19 @@ def main():
     clipboard_content = pyperclip.paste()
     print(clipboard_content)
     
-    # Start recording
-    read_aloud_stream(clipboard_content)
+    use_file = False
+    
+    if use_file:
+        read_aloud_file(clipboard_content)
+    else:
+        # Start recording
+        read_aloud_stream(clipboard_content)
 
-    # Start listening to keyboard events
-    with keyboard.Listener(on_press=on_press, on_release=on_release) as listener:
-        listener.join()  # Wait for the listener to finish
+        # Start listening to keyboard events
+        with keyboard.Listener(on_press=on_press, on_release=on_release) as listener:
+            listener.join()  # Wait for the listener to finish
 
-    subprocess.Popen(f"killall piper", shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+        subprocess.Popen(f"killall piper", shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
 
 if __name__ == "__main__":
     main()
